seenNotSeen button moved to MovieControllerCell with .toggle() utilized.  Cycles through both strings but won't stay on selected state
