+++ ios-sprint-challenge-movie-list-master(iOSPT5)/Movie List/Base.lproj/Main.storyboard
@@ -48,7 +48,7 @@
                                                     <rect key="frame" x="295" y="6" width="68" height="30.5"/>
                                                     <state key="normal" title=" Not Seen"/>
                                                     <connections>
-                                                        <action selector="seenNotSeenPressed:" destination="f82-4X-9sU" eventType="touchUpInside" id="jjt-y2-yae"/>
+                                                        <action selector="seenNotSeenPressed:" destination="CD6-Ve-sjh" eventType="touchUpInside" id="lr2-96-DEg"/>
                                                     </connections>
                                                 </button>
                                             </subviews>
@@ -64,6 +64,7 @@
                                         <connections>
                                             <outlet property="movieNameLabel" destination="fYu-Ty-J8U" id="1rP-qA-QEB"/>
                                             <outlet property="movieYearLabel" destination="W7l-wB-dXj" id="yKA-fu-vHy"/>
+                                            <outlet property="seenButton" destination="H62-4P-n6b" id="tH9-qH-gCx"/>
                                         </connections>
                                     </tableViewCell>
                                 </prototypes>
